Move set and clear user outside of _app
