+++ front/pages/_app.tsx
@@ -8,11 +8,8 @@ import "@app/styles/components.css";
 import { datadogLogs } from "@datadog/browser-logs";
 import type { NextPage } from "next";
 import type { AppProps } from "next/app";
-import { useRouter } from "next/router";
-import { useEffect } from "react";
 
 import RootLayout from "@app/components/app/RootLayout";
-import { useUser } from "@app/lib/swr/user";
 
 if (process.env.NEXT_PUBLIC_DATADOG_CLIENT_TOKEN) {
   datadogLogs.init({
@@ -40,38 +37,6 @@ type AppPropsWithLayout = AppProps & {
 export default function App({ Component, pageProps }: AppPropsWithLayout) {
   // Use the layout defined at the page level, if available.
   const getLayout = Component.getLayout ?? ((page) => page);
-
-  const router = useRouter();
-  const { wId } = router.query;
-  useEffect(() => {
-    const updateLoggerContext = () => {
-      if (wId) {
-        datadogLogs.setGlobalContext({
-          workspaceId: wId,
-        });
-      } else {
-        datadogLogs.setGlobalContext({});
-      }
-    };
-
-    updateLoggerContext();
-    router.events.on("routeChangeComplete", updateLoggerContext);
-    return () => {
-      router.events.off("routeChangeComplete", updateLoggerContext);
-    };
-  }, [wId, router.events]);
-
-  const { user } = useUser();
-  const userId = user?.sId;
-  useEffect(() => {
-    if (userId) {
-      datadogLogs.setUser({
-        id: userId,
-      });
-    } else {
-      datadogLogs.clearUser();
-    }
-  }, [userId]);
 
   return (
     <RootLayout>
